ATM card/requirements.txt

#================= required libraries in this project ==============
#================= run pip install -r requirements.tx ==============

opencv-python
pytesseract
pymongo
numpy
mysql-connector-python
